Move files to temp dir to avoid overwriting on async calls
Also changed return value for regression testing to "regression tests
passed"

+++ email/classes/EmailProcessor.php
@@ -27,14 +27,32 @@ class EmailProcessor
 		return true;
 	}
 
-	public static function processMessage($message, &$error) {
+	public static function processMessage($message, &$error, $deleteAttachments = false) {
 		if (EmailProcessor::simulate()) {
-			deleteAttachments($message['attachments']);
+			if ($deleteAttachments) { EmailProcessor::deleteAttachments($message['attachments']); };
 			echo trim(preg_replace('/\s+/', ' ', print_r($message, true)));
 			return;
 		}
 
-		deleteAttachments($message['attachments']);
+		try {
+			/*
+				$handler = EmailResponder::factory($message['to']);
+				$handler->process($message);
+
+				events@wycliffe-services.net
+				help@wycliffe-services.net
+				webservice@wycliffe-services.net
+
+				1) If email not recognized or is blank, send template.  Need template read from file
+				2) If template, convert to web service call and handle it from there.  Need body parser and mapping to web service call
+
+				file_put_contents('/var/www/email/output.html', '<pre>' .print_r($message, true) . '</pre>');
+			*/
+		} catch (Exception $e) {
+		
+		}
+		
+		if ($deleteAttachments) { EmailProcessor::deleteAttachments($message['attachments']); }
 		echo '<pre>' .print_r($message, true) . '</pre>';
 	}
 
@@ -112,6 +130,7 @@ class EmailProcessor
 	}
 	private static function initAttachments($mail, &$struct, $buffer, &$message, &$error) {
 		$message['attachments'] = array();
+		$baseDir = util::createTempDir();
 
 		while (!empty($struct)) {
 			$info = EmailProcessor::getInfo($mail, array_shift($struct));
@@ -121,7 +140,7 @@ class EmailProcessor
 				
 				$temp = EmailProcessor::getInfo($mail, array_shift($struct));
 				$info['disposition-filename'] = $temp['headers']['subject'] . '.eml';
-				EmailProcessor::processAttachment($buffer, $info, $message);
+				EmailProcessor::processAttachment($buffer, $info, $baseDir, $message);
 				
 				if (count($struct) == 0) { return; }
 				while (util::startsWith($struct[0], $key)) {
@@ -130,12 +149,15 @@ class EmailProcessor
 				continue;
 			}
 			if ($info['content-disposition'] == 'attachment') {
-				EmailProcessor::processAttachment($buffer, $info, $message);
+				EmailProcessor::processAttachment($buffer, $info, $baseDir, $message);
 				continue;
 			}
 			$error = 'Unexpected info: ' . print_r($info, true);
+			util::deltree($baseDir);
 			return false;
 		}
+		
+		if (count($message['attachments']) == 0) { rmdir($baseDir); }
 		return true;
 	}
 
@@ -148,8 +170,8 @@ class EmailProcessor
 		return mailparse_msg_get_part_data($section);
 	}
 
-	private static function processAttachment($buffer, $info, &$message) {
-		$newPath = sys_get_temp_dir() . DIRECTORY_SEPARATOR . $info['disposition-filename'];
+	private static function processAttachment($buffer, $info, $baseDir, &$message) {
+		$newPath = $baseDir . $info['disposition-filename'];
 		$message['attachments'][] = $newPath;
 		if ($info['transfer-encoding'] == 'base64') { 
 			file_put_contents($newPath, base64_decode(EmailProcessor::getData($buffer, $info)));
@@ -168,6 +190,14 @@ class EmailProcessor
 		$simulate = isset($_GET['simulate']) ? $_GET['simulate'] : (isset($_POST['simulate']) ? $_POST['simulate'] : 0);
 		$simulate = filter_var($simulate, FILTER_VALIDATE_INT, array('filter'=>FILTER_VALIDATE_INT, 'options'=>array("min_range"=>0, "max_range"=>1)));
 		return $simulate == 1;
+	}
+	
+	private static function deleteAttachments($attachments) {
+		if (count($attachments) == 0) { return; }
+		
+		$arr = explode(DIRECTORY_SEPARATOR, $attachments[0]);
+		array_pop($arr);
+		util::delTree(implode(DIRECTORY_SEPARATOR, $arr));
 	}
 }
 ?>
